requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ansicon==1.89.0
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
blessed==1.20.0
carla==0.9.13
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
croniter==1.3.8
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.0
dnspython==2.3.0
docker-pycreds==0.4.0
docopt==0.6.2
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.88.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.3.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
imageio==2.25.1
importlib-resources==5.12.0
iniconfig==2.0.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
kiwisolver==1.4.4
lightning==2.0.1
lightning-cloud==0.5.32
lightning-utilities==0.8.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
multidict==6.0.4
networkx==2.6.2
numpy==1.20.3
opencv-python==4.7.0.68
ordered-set==4.1.0
orjson==3.8.8
packaging==23.0
pandas==1.3.2
pathtools==0.1.2
Pillow==9.4.0
pipreqs==0.4.11
pluggy==1.0.0
protobuf==4.22.0
psutil==5.9.4
py-trees==2.2.3
pydantic==1.10.7
pydot==1.4.2
pygame==2.1.3
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.0
pytorch-nlp==0.5.0
pytz==2022.7.1
PyYAML==6.0
readchar==4.0.5
requests==2.28.2
rfc3986==1.5.0
rich==13.3.2
scikit-learn==0.24.2
scipy==1.10.1
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.22.0
starsessions==1.3.0
threadpoolctl==3.1.0
tk==0.1.0
tomli==2.0.1
torch==1.13.1+cu116
torch-cluster==1.6.0+pt113cu116
torch-geometric==2.2.0
torch-scatter==2.1.0+pt113cu116
torch-sparse==0.6.16+pt113cu116
torch-spline-conv==1.2.1+pt113cu116
torchaudio==0.13.1+cu116
torchmetrics==0.11.4
torchvision==0.14.1+cu116
tqdm==4.62.2
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.21.1
wandb==0.14.0
watchfiles==0.18.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
yarg==0.1.9
yarl==1.8.2
zipp==3.14.0